pyproject.toml
[tool.poetry]
name = "suap-bot"
version = "0.1.0"
description = ""
authors = ["rafaeldss <rafaeldss199916@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-telegram-bot = "^13.14"
PyJWT = "^2.4.0"
Flask-SQLAlchemy = "^3.0.1"
Flask-Admin = "^1.6.0"
Flask-Migrate = "^3.1.0"
Flask-WTF = "^1.0.1"
requests = "^2.28.1"
pydantic = "^1.9.2"
celery = "^5.2.7"
PyAMQP = "^0.1.0"
flower = "^1.2.0"

[tool.poetry.dev-dependencies]
honcho = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
